backend/requirements.txt
fastapi
uvicorn
pydantic
youtube-transcript-api
faiss-cpu
sentence-transformers
openai
python-dotenv
